feat: abstract utility functions of useUrgency hook into a separate urgencyUtils file to keep the code readable and scalable
